requirements.txt
django
djangorestframework
django-cors-headers
pandas
numpy
matplotlib